Suppress more warnings and repod with .def files

+++ Source/KSCrash/Source/KSCrash/swift/Basic/Demangler.cpp
@@ -1180,7 +1180,7 @@ NodePointer Demangler::popFunctionParamLabels(NodePointer Type) {
         return LabelList;
     
     bool hasLabels = false;
-    for (unsigned long i = 0, n = ParameterType->getIndex(); i != n; ++i) {
+    for (unsigned long long i = 0, n = ParameterType->getIndex(); i != n; ++i) {
         auto Label = getLabel(Tuple, i);
         
         if (!Label)
